+++ FireworkFlutterSDKSample/lib/models/app_language_info.g.dart
@@ -8,12 +8,12 @@ part of 'app_language_info.dart';
 
 AppLanguageInfo _$AppLanguageInfoFromJson(Map<String, dynamic> json) =>
     AppLanguageInfo(
-      language: json['language'] as String?,
+      languageCode: json['languageCode'] as String?,
       displayName: json['displayName'] as String?,
     );
 
 Map<String, dynamic> _$AppLanguageInfoToJson(AppLanguageInfo instance) =>
     <String, dynamic>{
-      'language': instance.language,
+      'languageCode': instance.languageCode,
       'displayName': instance.displayName,
     };