+++ FireworkFlutterSDKSample/lib/utils/fw_example_logger_util.dart
@@ -1,10 +1,15 @@
 import 'dart:developer' as developer;
 
 class FWExampleLoggerUtil {
+  static final messageList = <String>[];
   static void log(
     String message, {
     String name = 'com.loopnow.fondor_example',
+    bool shouldCache = false,
   }) {
     developer.log(message, name: name);
+    if (shouldCache) {
+      messageList.add(message);
+    }
   }
 }